+++ example/src/pages/index/components/Footer/index.jsx
@@ -4,7 +4,6 @@ import { connect } from 'nuomi';
 class Footer extends Component {
   render() {
     const { total, count, dispatch } = this.props;
-    console.log(111)
     return <div onClick={() => {
       dispatch('@update', {
         count: count + 1