+++ packages/nuomi/src/hooks/index.js
@@ -1,2 +1,3 @@
 export { default as useConnect } from './useConnect';
 export { default as useNuomi } from './useNuomi';
+export { default as useLocation } from './useLocation';