+++ packages/nuomi/src/hooks/useNuomi.js
@@ -5,7 +5,7 @@ import { NuomiContext } from '../components/Context';
 function useNuomi() {
   const { nuomi } = useContext(NuomiContext);
   invariant(nuomi, '不允许在 <Route>、<Nuomi>、<NuomiRoute> 外部使用 useNuomi');
-  return [nuomi];
+  return nuomi;
 }
 
 export default useNuomi;