+++ packages/nuomi/types/component.d.ts
@@ -46,3 +46,5 @@ export class NavLink<P extends NavLinkProps> extends PureComponent<P, any> {
 export const connect: Connect;
 
 export function withNuomi<P extends { nuomi: Context }>(Component: ComponentType<P>): ComponentType<P>;
+
+export function withRouter<P extends { location: Location }>(Component: ComponentType<P>): ComponentType<P>;