+++ packages/nuomi/types/hooks.d.ts
@@ -1,7 +1,10 @@
 import { Context } from './props';
 import { ConnectGetState } from './connect';
 import { NuomiStoreDispatch } from './store';
+import { Location } from './router';
 
 export function useConnect<S = any, G = any>(getState?: ConnectGetState<S, G>): [S & { getter: G }, NuomiStoreDispatch];
 
-export function useNuomi<S = any, G = any>(): [Context<S, G>];
+export function useNuomi<S = any, G = any>(): Context<S, G>;
+
+export function useLocation(): Location;